Rewrote some tests to run for all eBPF programs.
Formatting changes

Rewrote more tests to run for all eBPF programs.

Modified certain tests to run for all types of eBPF programs.

Modified specific netsh_test to include all eBPF program types.

Added a scope exit to ensure cleanup during fault injection.

Reordered cleanup in test helper destructor.
